+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.2 (2026-06-27)
+
+- Fix crash when `server_url` contains query params.
+
 ## 1.1.1 (2023-08-13)
 
 - Fix error in error message (sic) when hackney is missing